requirements.txt
# --- FastAPI + server ---
fastapi==0.115.0
uvicorn[standard]==0.30.0

# --- Docling core ---
docling==0.9.2
docling-core==0.9.2

# --- PDF & conversion dependencies ---
lxml==5.2.1
python-multipart==0.0.12

# --- OCR engines (Docling uses whichever is available) ---
pytesseract==0.3.13
rapidocr-onnxruntime==1.3.13
onnxruntime==1.19.0

# --- EasyOCR + Torch (CPU build only) ---
easyocr==1.7.1
torch==2.3.0 --extra-index-url https://download.pytorch.org/whl/cpu

# --- SSH / SFTP for your remote upload ---
paramiko==3.4.0

# --- Markdown / utility ---
markdownify==0.12.1

# --- Other dependencies used by Docling in pipelines ---
numpy==1.26.4
Pillow==10.3.0

# --- Load .env files ---
python-dotenv>=1.0.0